JAudio: Rename inconsistent headers
Now these are named the same as their associated source files.
