+++ include/IntroGameSection.h
@@ -10,7 +10,7 @@
 #include "EffectMgr.h"
 #include "MoviePlayer.h"
 #include "Graphics.h"
-#include "jaudio/PikiScene.h"
+#include "jaudio/piki_scene.h"
 #include "ModeState.h"
 
 /**
