+++ src/jaudio/app_inter.c
@@ -1,7 +1,7 @@
 #include "jaudio/app_inter.h"
 
-#include "jaudio/PikiScene.h"
-#include "jaudio/PikiBgm.h"
+#include "jaudio/piki_scene.h"
+#include "jaudio/piki_bgm.h"
 #include "jaudio/hvqm_play.h"
 
 static BOOL demo_fade_flag;
