+++ src/jaudio/piki_bgm.c
@@ -1,10 +1,10 @@
-#include "jaudio/PikiBgm.h"
+#include "jaudio/piki_bgm.h"
 
 #include "jaudio/playercall.h"
 #include "jaudio/file_seq.h"
 #include "jaudio/seqsetup.h"
 #include "jaudio/interface.h"
-#include "jaudio/PikiScene.h"
+#include "jaudio/piki_scene.h"
 #include "jaudio/jammain_2.h"
 #include "Dolphin/os.h"
