+++ src/jaudio/piki_player.c
@@ -1,8 +1,8 @@
-#include "jaudio/PikiPlayer.h"
-#include "jaudio/PikiBgm.h"
+#include "jaudio/piki_player.h"
+#include "jaudio/piki_bgm.h"
 #include "jaudio/jammain_2.h"
-#include "jaudio/PikiDemo.h"
-#include "jaudio/PikiScene.h"
+#include "jaudio/pikidemo.h"
+#include "jaudio/piki_scene.h"
 #include "jaudio/verysimple.h"
 #include "jaudio/ja_calc.h"
 #include "jaudio/random.h"
