+++ src/jaudio/piki_scene.c
@@ -1,9 +1,9 @@
-#include "jaudio/PikiScene.h"
+#include "jaudio/piki_scene.h"
 
-#include "jaudio/PikiBgm.h"
-#include "jaudio/PikiInter.h"
-#include "jaudio/PikiDemo.h"
-#include "jaudio/PikiPlayer.h"
+#include "jaudio/piki_bgm.h"
+#include "jaudio/pikiinter.h"
+#include "jaudio/pikidemo.h"
+#include "jaudio/piki_player.h"
 #include "jaudio/waveread.h"
 #include "jaudio/aramcall.h"
 #include "jaudio/interface.h"
