+++ src/jaudio/pikidemo.c
@@ -1,9 +1,9 @@
-#include "jaudio/PikiDemo.h"
+#include "jaudio/pikidemo.h"
 #include "jaudio/jammain_2.h"
 #include "jaudio/cmdqueue.h"
-#include "jaudio/PikiScene.h"
-#include "jaudio/PikiPlayer.h"
-#include "jaudio/PikiBgm.h"
+#include "jaudio/piki_scene.h"
+#include "jaudio/piki_player.h"
+#include "jaudio/piki_bgm.h"
 #include "jaudio/waveread.h"
 #include "jaudio/interface.h"
 #include "jaudio/verysimple.h"
