+++ src/jaudio/verysimple.c
@@ -8,15 +8,15 @@
 #include "jaudio/jammain_2.h"
 #include "jaudio/cmdqueue.h"
 #include "jaudio/syncstream.h"
-#include "jaudio/PikiBgm.h"
-#include "jaudio/PikiPlayer.h"
-#include "jaudio/PikiDemo.h"
-#include "jaudio/PikiScene.h"
+#include "jaudio/piki_bgm.h"
+#include "jaudio/piki_player.h"
+#include "jaudio/pikidemo.h"
+#include "jaudio/piki_scene.h"
 #include "jaudio/waveread.h"
 #include "jaudio/aictrl.h"
 #include "jaudio/cmdstack.h"
 #include "jaudio/fat.h"
-#include "jaudio/PikiInter.h"
+#include "jaudio/pikiinter.h"
 #include "jaudio/fxinterface.h"
 #include "jaudio/bankloader.h"
 #include "jaudio/seqsetup.h"
